==> 05 Final Project/01 Code Books/01 Tickers.ipynb <==
# Tickers

Tracking Tickers by Stock Exchange

## Import Modules

import pandas as pd
import os

## Import Data

home_path = os.getcwd()
home_path += "\\"

data_path = home_path + "02 Data Files\\"

print(data_path)

nyse = pd.read_csv(data_path+"01 Tickers\\"+"nyse"+"_tickers.txt", header=None)
nasdaq = pd.read_csv(data_path+"01 Tickers\\"+"nasdaq"+"_tickers.txt", header=None)
amex = pd.read_csv(data_path+"01 Tickers\\"+"amex"+"_tickers.txt", header=None)

## Clean Data

nyse = nyse.rename(columns={0 : "Ticker"})
nyse["Exchange"] = "NYSE"
nyse

nasdaq = nasdaq.rename(columns={0 : "Ticker"})
nasdaq["Exchange"] = "NASDAQ"
nasdaq

amex = amex.rename(columns={0 : "Ticker"})
amex["Exchange"] = "AMEX"
amex

## Check for Duplicates

tickers = set()

for df in [nyse, nasdaq, amex]:
    tickers.update(df["Ticker"])

tickers = pd.DataFrame(list(tickers), columns=["Ticker"])

def indf(key, df):
    return key in set(df["Ticker"])

name_map = {"nyse": "NYSE", "nasdaq": "NASDAQ", "amex": "AMEX"}

for df_name, df in zip(name_map.keys(), [nyse, nasdaq, amex]):
    tickers[name_map[df_name]] = tickers["Ticker"].apply(lambda x: indf(x, df))

tickers["Number of Exchanges"] = tickers[['NYSE', "NASDAQ", "AMEX"]].sum(axis=1)

tickers = tickers.sort_values(by="Number of Exchanges", ascending=False).reset_index(drop=True)

remove_tickers = list(tickers[tickers["Number of Exchanges"]>1]["Ticker"])

print(remove_tickers)

print(f"\nThe number of tickers being removed: {len(remove_tickers)} or {round(len(remove_tickers)/len(tickers)*100,2)}% of tickers")

tickers = pd.DataFrame(tickers[tickers["Number of Exchanges"]==1]["Ticker"].reset_index(drop=True))

tickers.to_csv(data_path+"01 Tickers\\"+"final_tickers.csv", index=False)

tickers
