==> 05 Final Project/01 Code Books/03 Data Integration.ipynb <==
# Scraping Capitol Trades


## Import Modules

import pandas as pd  # For handling data in DataFrame
from selenium import webdriver  # For controlling the web browser and interacting with HTML content
from selenium.webdriver.chrome.service import Service  # For setting up ChromeDriver as a service
from selenium.webdriver.common.by import By  # To locate HTML elements by XPath
from selenium.webdriver.chrome.options import Options  # For handling Chrome options
from webdriver_manager.chrome import ChromeDriverManager  # For automatic ChromeDriver installation
import time  # For handling wait times
import random  # For randomizing scroll and wait times
import os
import glob #finding all of the subset files

### Set Up Paths

home_path = os.getcwd()
home_path = os.path.dirname(home_path)
home_path += "\\"

data_path = home_path + "02 Data Files\\"

folder_path = data_path + "02 Trade Scraping\\"

print(data_path)

## Import Data and Combine Data

# Find all CSV files that start with "transaction_log_scraped_subset"
csv_files = glob.glob(os.path.join(folder_path, "transaction_log_scraped_subset*.csv"))

# Read and combine all CSV files into a single DataFrame
df = pd.concat([pd.read_csv(file) for file in csv_files], ignore_index=True)

df

## Clean Data

df["Date_As_DateTime"] = pd.to_datetime(df["Date"])

df = df[df["Date_As_DateTime"].dt.year.isin([2023, 2024])] #Filter to just 2023 and 2024

df

### Cleaning Tickers

Import Ticker List

tickers = pd.read_csv(data_path+"01 Tickers\\"+"final_tickers.csv")

tickers = set(tickers["Ticker"])

tickers

Define inset()

This will be used to check if the ticker in the df is in the set of valid tickers

def inset(key, set=tickers):
    return key in set

Evaluate Tickers and Drop Invalid Tickers

df["Include"] = df["Ticker"].apply(lambda key: inset(key))

df = df[df["Include"]==True].reset_index(drop=True)

df

### Save Cleaned Data to CSV

df.to_csv(data_path+"02 Trade Scraping\\"+"transition_log_scraped.csv")

## Split Data by Transaction Type

buydf = df[df["Buy"]==True][['Name',"Party",'Chamber', 'State', 'Ticker', 'Date']].reset_index(drop=True)

buydf

selldf = df[df["Sell"]==True][['Name',"Party",'Chamber', 'State', 'Ticker', 'Date']].reset_index(drop=True)

selldf

## Integrate Data by Matching Transactions

merged_df = buydf.merge(selldf, on=["Name", "Ticker", "Party", "Chamber", "State"], suffixes=("_buy", "_sell"))

merged_df

matched_df = merged_df.groupby(["Name", "Ticker", "Party", "Chamber", "State"]).agg(
    Earliest_Buy_Date=("Date_buy", "min"),
    Latest_Sell_Date=("Date_sell", "max")
).reset_index()

matched_df = matched_df.rename(columns={"Earliest_Buy_Date" : "Buy Date", "Latest_Sell_Date" : "Sell Date"})

matched_df["Buy Date"] = pd.to_datetime(matched_df["Buy Date"], errors="coerce")
matched_df["Sell Date"] = pd.to_datetime(matched_df["Sell Date"], errors="coerce")

#Filter out when Earliest Buy Date is After Latest Sell Date , since stocks could have been bought before sample range
matched_df = matched_df[matched_df["Sell Date"] >= matched_df["Buy Date"]]

matched_df["Buy Date"] = matched_df["Buy Date"].dt.strftime("%Y-%#m-%#d")
matched_df["Sell Date"] = matched_df["Sell Date"].dt.strftime("%Y-%#m-%#d")

matched_df

## Output Matched and Integrated Data

matched_df.to_csv(data_path+"03 Data Integration\\"+"transactions_matched.csv", index=False)
